Store Integers, not Floats
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@5649 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/MPRestoreItemList.java
@@ -162,7 +162,7 @@ public abstract class MPRestoreItemList
 			this.purchaseCost = purchaseCost;
 			this.isCombatUsable = isCombatUsable;
 
-			MPRestoreItemList.manaRestoredByName.put( itemName, new Float( this.manaPerUse ) );
+			MPRestoreItemList.manaRestoredByName.put( itemName, new Integer( this.manaPerUse ) );
 
 			if ( ItemDatabase.contains( itemName ) )
 			{
